+++ bin/archives/ch01.md
@@ -1,0 +1,18 @@
+# 1장 기본 알고리즘
+
+## 알고리즘 정의
+
+- 문제 해결을 위한 처리 절차
+- 정당성: 올바른 결과 반환
+- 정지성: 정해진 시간 내에 반드시 종료
+
+## 구조적 프로그래밍
+
+- 비구조적 프로그래밍
+  - 하나의 연속된 덩어리에 모든 코드를 넣음
+- 구조적 프로그래밍
+  - 절차적(procedural) 프로그래밍(프로시저는 함수를 의미)
+    - 프로그래밍의 흐름이 순차적임
+    - 조건 분기로 흐름을 제어 가능
+    - 반복문으로 흐름을 반복 수행 가능
+    - 특정 기능을 모듈화 가능 => 반복되는 모듈을 재사용 가능한 프로시저 단위로 나누고 필요할때마다 해당 함수 호출
