+++ bin/archives/ch02.md
@@ -1,0 +1,26 @@
+# 2장 기본 자료구조
+
+## 자료구조 정의
+
+- 자료를 효율적으로 이용하기 위한 데이터 저장 방법
+- 데이터 단위는 데이터를 구성하는 한 덩어리
+
+## 2-1 배열
+
+- 같은 타입의 요소가 연속되어 있는 것
+- 각 요소는 순서가 있음
+- 메서드 매개변수로 배열 사용시 배열 본체에 대해 참조한다. 메서드에서 받은 배열은 넘겨받은 배열 본체를 직접 참조한다.
+- 배열 요소 역순 정렬
+  - 중간을 기준으로 맨 앞과 맨 뒤 요소를 교환
+  - 교환횟수 = 요소갯수 / 2
+  - 왼쪽은 0부터 1씩 증가, 오른쪽은 n부터 1씩 감소
+  - 두 값을 교환할 땐 임시변수 사용
+- 기수 변환
+  - 정수값을 x, 진수를 r로 놓고 배열에 변환한 수를 저장하면 결과값의 역순으로 저장됨
+- 소수 나열
+  - 첫번째) 2부터 n-1까지 반복문 돌면서 그 사이의 어떤 정수로도 나누어 떨어지지 않으면 소수
+  - 개선 1) 2는 소수, 짝수는 다 소수이므로 대상은 홀수만, 소수를 배열에 저장후 배열 안의 수로 나누어 떨어지는 수가 아니면 소수
+  - 개선 2) 2, 3은 소수, n의 제곱근 이하 어떤 소수로도 나누어 떨어지지 않으면 소수
+- 클래스
+  - 여러 형의 요소를 조합해 만든 자료구조
+  - 클래스형 변수는 실체 인스턴스를 참조한다.
